pyproject.toml
[project]
name = "multi-agent-content-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.0",
]
